Move BaseForm in Abstracts to top level
