+++ src/app/employee/employee-form/bio-form/bio-form.component.ts
@@ -1,10 +1,10 @@
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core'
 import { AbstractControl, FormBuilder, FormGroup, Validators } from '@angular/forms'
 
+import { BaseForm } from '../../../abstracts/base-form.class'
 import { ETHNICITY, IEmployeeBio } from '../../../models/employee.interface'
 import { DateValidators } from '../../../utils/date.validators'
 import { hasChanged } from '../../../utils/functions'
-import { BaseForm } from '../../abstracts/base-form.class'
 
 @Component({
   selector: 'tcp-bio-form',
